Update Project_compare.Rmd
I was working on couple to-do lists for the code:
(1) Whether to include variable "addr_state" or not
I lean towards removing it because it doesn't seem to influence the outcome. Although I am not sure if Lending Club would have budget of the loan for borrowers in each state. The more that I read, I think zipcode should definitely be removed because it could a proxy of status and may not be appropriate for decision making.

(2) Remove features that contain one unique value + NA
I kind of did it manually, any better/neater solution is welcome.

(3) Check the dictionary of variables
I uploaded an excel sheet to Onedrive named "Dictionary_check" to complete this task, and I assigned keep_score to each variable, 1 means to keep, 0 means to not keep. Basically, Sebastian and I don't have conflict. I created a column for Monica, feel free to use it ! 👍 Also, "grade" seems to contain "subgrade" and "total_acc" seems to contain "open_acc", should we remove one of each?

(4) Convert features to the appropriate categories
I found that the format of variable earliest_cr_line seems messy. I haven't figured out how to convert it into numerical.